// src/supabase.d.ts
export interface definitions {
  subscriptions: {
    /**
     * Format: uuid
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    created_at?: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    updated_at?: string;
    /** Format: character varying */
    name: string;
    /** Format: text */
    description?: string;
    /** Format: date */
    start_date: string;
    /** Format: date */
    end_date?: string;
    /** Format: character varying */
    fee_recurrence: string;
    /** Format: double precision */
    fee: number;
    /** Format: character varying */
    category?: string;
  };
}